Task 2 - Created Borrower Class
Create a borrower class that tracks the information of when a book is checked out or returned by a user. This is done through a method that adds the book title assigned to the name to the borrowed books array and a method that removes the book when it is returned by the user.