Task 3 - Created Library Class
Create a library class that stores the list of books and the list of borrowers for the books. Two methods are added, one to add a book to the library and one two pull the information for each book in the library and print an updated list of all the details of the books.